+++ internal/linkcheck/status.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/axllent/mailpit/config"
-	"github.com/axllent/mailpit/utils/logger"
+	"github.com/axllent/mailpit/internal/logger"
 )
 
 func getHTTPStatuses(links []string, followRedirects bool) []Link {